% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloridaBayShp.R
\name{fb}
\alias{fb}
\title{Florida Bay shapefile with subregions}
\format{
a SpatialPolygonsDataFrame
}
\description{
A SpatialPolygonsDataFrame. proj4string: [+proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
}
\examples{
summary(fb)

\dontrun{
### code used to generate object 
fb <- terra::vect("C:/ESRIDATA/Florida_Bay/subregions", "SFL_NNC") 
fb <- fb[grepl(x = fb$SEGMENT_NA, pattern = "Florida Bay|Manatee Bay"), ]
fb <- terra::project(fb, terra::crs(terra::rast(fireHydro::edenDEM), proj = TRUE))
# save("fb", file =  paste0(here::here(), "/data/fb.RData"))

}
}
