% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feetToCm.R
\name{feetToCm}
\alias{feetToCm}
\title{Convert feet to centimeters}
\usage{
feetToCm(ft)
}
\arguments{
\item{ft}{input value with units of feet.}
}
\value{
A numeric value with units of centimeters
}
\description{
Converts feet to centimeters
}
\examples{
feetToCm(0.18)

 
}
