% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEDEN.R
\name{getEDENbyROI}
\alias{getEDENbyROI}
\title{getEDENbyROI}
\usage{
getEDENbyROI(targetLocations, targetLocationNames, dateRange, needEDEN)
}
\arguments{
\item{targetLocations}{locations to extract data from; must be class SpatialPointsDataFrame or SpatialPolygonsDataFrame}

\item{targetLocationNames}{option to specify the name of target locations (e.g., pts$gage)}

\item{dateRange}{Date range to download EDEN data. Must be in \%Y\%m\%d (e.g., "20201231") format. This can be a character vector.}

\item{needEDEN}{The name of the list of EDEN data generated from, e.g., lapply(dateRange, fireHydro::getEDEN) or a rasterStack as generated by fireHydro::getAnnualEden(2020). Specifically, object is a list of lists with (1) a date (YYYYMMDD) and (2) EDEN data (must be a rasterStack; in `getEDEN()` use argument `returnType = 'raster'`).}
}
\value{
a dataframe formatted identically to the traceDataLong output from extractSimData()
}
\description{
This function extracts EDEN data for spatial locations of interest. This function can take a very long time depending on the number of regions-of-interest and their size, and if EDEN data is downloaded within the function call. NOTE: This is based on availability of data on the home page of EDEN. If data aren't there, they aren't accessible for this code.
}
\examples{
 

\dontrun{

IR_pattern <- "WCA-1"
dateList <- format(seq.Date(from = as.Date("20200101", format = "\%Y\%m\%d"), 
            to = as.Date("20200109", format = "\%Y\%m\%d"), by = "day"), format = "\%Y\%m\%d")

### pull EDEN data separately  so it's retained in the global environment
EDEN_data <- lapply(dateList, fireHydro::getEDEN)
EDEN_by_IR <- getEDENbyROI(targetLocations = ROIs,
                           targetLocationNames = targetLocationNames,
                           dateRange = dateList, 
                           needEDEN = EDEN_data)
              
}

}
