% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuantiles.R
\name{getQuantiles}
\alias{getQuantiles}
\title{Summarizes EverForecast simulations.}
\usage{
getQuantiles(
  traceDataLong,
  ROI_names = paste0(unique(traceDataLong$name), collapse = "|"),
  quantiles = c(0.1, 0.25, 0.5, 0.75, 0.9)
)
}
\arguments{
\item{traceDataLong}{Lightly processed EverForecast simulation data, found in the traceDataLong object returned by RSM::extractSimData(). RSM::extractSimData() needs to be run for this function to be used.}

\item{ROI_names}{grep-style character vector used to select a subset of the targetLocations in the input in the traceDataLong object.}

\item{quantiles}{percentiles to include in output. For example, 0.5 would return the 50th percentile (median value from x simulations) for each day in the EverForecast simulations and for each region of interest identified in ROI_names}
}
\value{
list \code{getQuantiles} returns a long dataframe with one row per date, region of interest, and quantile
}
\description{
Generates a dataframe of daily EverForecast quantiles for each point/region of interest (where population is the mean water depth on a given day in n EverForecast simulations).
}
\examples{
getQuantiles
\dontrun{
}

}
