% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroperiod.R
\name{hydroperiod}
\alias{hydroperiod}
\title{Quantifies hydroperiod (continuous or discontinuous)}
\usage{
hydroperiod(x, threshold = 0, continuous = TRUE, percent = FALSE)
}
\arguments{
\item{threshold}{threshold value for determining inundation. Note: values equal to or greater than the threshold count as inundated.}

\item{continuous}{if TRUE, the quantity calculated is continuous hydroperiod (largest run of consecutive days inundated). If `FALSE`, the function returns discontinuous hydroperiod (total days inundated in record)}

\item{percent}{if TRUE, returned value is expressed as a fraction of all non-NA observations. This may not be meaningful in all cases (e.g., continuous hydroperiod)}

\item{data}{input dataset (e.g., one year of stage data for a point/pixel)}
}
\value{
numeric \code{hydroperiod} returns a numeric value with time units equivalent to the time increments in input data
}
\description{
Calculates the duration of inundation, either continuous or discontinuous
}
\examples{
dat <- 1:100
hydroperiod(data = dat, threshold = 75, continuous = TRUE)
\dontrun{
}

 
}
