% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSims.R
\name{loadSims}
\alias{loadSims}
\title{Loads and pre-processes EverForecast simulations}
\usage{
loadSims(directory = file.choose())
}
\arguments{
\item{directory}{the address of an individual file in a folder of EverForecast .nc files. The default action is <code>file.choose()</code>; a browser menu appears so the user can select the the desired directory by identifying a single .nc file in the folder of netCDFs.}
}
\value{
A rasterStack object with units feet NGVD29
}
\description{
Loads netCDF files and converts units from cm depth to feet NGVD29. highly recommended to set raster options as in the example section.
}
\examples{
loadSims
\dontrun{
### recommended raster options to save to disk rather than work in 
### memory
rasterOptions(progress = "text", timer = TRUE, overwrite = TRUE, 
todisk=TRUE, # critical to write temp files to disk rather than store in RAM
tmpdir = "G:/data/raster/temp/delete", # critical to specify a temp folder
tmptime = 4)

 }
 
}
