% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxidation.R
\name{oxidation}
\alias{oxidation}
\title{Quantifies soil oxidation}
\usage{
oxidation(data, threshold = 0)
}
\arguments{
\item{data}{input dataset (e.g., one year of stage data for a point/pixel)}

\item{threshold}{threshold is the value *below which* depths are accumulated. Values equal to the threshold are treated as inundated.}
}
\value{
numeric \code{oxidation} returns a numeric value with time units equivalent to the time increments in input data
}
\description{
Calculates soil oxidation performance measure. Accumulates length units when stages are below a threshold value.
}
\examples{
dat <- 1:100
oxidation(data = dat, threshold = 75)
\dontrun{
}

 
}
