% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPM.R
\name{plotPM}
\alias{plotPM}
\title{Plot extracted EverForecast and EDEN data}
\usage{
plotPM(quantileData, EDENdata, daysToPlot = 60)
}
\arguments{
\item{quantileData}{EverForecast quantile dataframe, as returned by `getQuantiles()`}

\item{EDENdata}{EDEN quantile dataframe, as returned by `getQuantiles()`}

\item{daysToPlot}{days of EDEN data to plot prior to EverForecast model run date}
}
\value{
A ggplot object
}
\description{
A ggplot wrapper for easily plotting extracted EDEN data and EverForecast output aggregated by `getQuantiles`
}
