% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_efc.R
\name{read.efc}
\alias{read.efc}
\title{Read loaded EverForecast data saved to disk}
\usage{
read.efc(directory = "G:/data/ever4cast/20220501/output/")
}
\arguments{
\item{directory}{character. Input directory}
}
\value{
list of SpatRasters
}
\description{
Load an EverForecast list saved with write.efc. Returns a list of SpatRasters.
}
\examples{

\dontrun{
everforecastDirectory <- "C:/path/to/data"
dat1 <- loadSims(directory = everforecastDirectory)

write.efc(x = dat1, directory = "G:/data/ever4cast/20220501/output/")
newdat <- read.efc(directory = "G:/data/ever4cast/20220501/output/")

newdat

}

}
