% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_efc.R
\name{write.efc}
\alias{write.efc}
\title{Write EverForecast data to disk}
\usage{
write.efc(x, directory = "G:/data/ever4cast/20220501/output/")
}
\arguments{
\item{x}{list of EverForecast SpatRaster data sets}

\item{directory}{character. Output directory; this is not a filename. An individual file is saved for each element in the EverForecast list}
}
\value{
none
}
\description{
Write a list of EverForecast data to a file folder
}
\examples{

\dontrun{
everforecastDirectory <- "C:/path/to/data"
dat1 <- loadSims(directory = everforecastDirectory)

write.efc(x = dat1, directory = "G:/data/ever4cast/20220501/output/")
newdat <- read.efc(directory = "G:/data/ever4cast/20220501/output/")

newdat

}

}
