% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColors.R
\name{getColors}
\alias{getColors}
\title{Generate a three-color gradient palette}
\usage{
getColors(
  uniqueValues = c(-2:4)/10,
  binSize = 0.025,
  colorGradient = c("red", "white", "blue"),
  threshold = 0
)
}
\arguments{
\item{uniqueValues}{input values (or vector with max/min)}

\item{binSize}{size of increment between color gradations}

\item{colorGradient}{vector with three colors that form the gradient}

\item{threshold}{value of gradient transition. This value will be the middle color listed in `colorGradient`}
}
\value{
A list with color codes and breaks
}
\description{
Converts feet to centimeters
}
\examples{
getColors(uniqueValues = range(-10:10), binSize = 5)
getColors(uniqueValues = range(-10:10), binSize = 1)

}
